Refactor: relocate extensions and routes to common directory and configure common templates path.
