+++ app/common/templates/background.html
